Create Dying Animation for player
When our player dies, we want to use tweens to make a death animation. This animation should rotate the player around itself multiple times during the duration, along with fading the player out to transparency. At the end of the tween, an event should be fired to let other areas in our code know that the death animation is over. This will require a bit of digging in the code to figure out how to get this to work.